requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.5
langchain-chroma>=0.1.2
langchain-huggingface>=0.0.3
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0
pypdf>=4.2.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
pandas>=2.2.0
unstructured>=0.14.0
tiktoken>=0.7.0
httpx>=0.27.0